booking-service/requirements.txt
Django==4.2
djangorestframework==3.14
pika==1.3.2
requests==2.31.0